requirements.txt
fhir.resources>=7.1.0
matplotlib>=3.7.0
